phase1046d-revived-wae-site-fixes-wave3(engine+samples+.clang-tidy)
Wave 3: 56 sites (chunk-3 ui/world/game/asset: 48 + chunk-1/2
recheck leftovers: 8) via 2 parallel developer agents on disjoint
file sets; verified centrally.

  - WasapiAudioBackend: 14x lock_guard -> scoped_lock; 1x contains;
    operator& overload KEPT with NOLINTNEXTLINE(google-runtime-
    operator) — COM out-param idiom (&ptr must yield T** for the
    WASAPI factory calls at lines 138-185).
  - ECS Scheduler.hpp: 3x contains + ranges::sort; ArchetypeWorld:
    widening cast; Obb.hpp: simplify-boolean-expr with the negated
    form kept to preserve exact NaN semantics.
  - test_scheduler: 8x emplace_back; NullAudioBackend: contains.
  - ui/game/asset tests: 4x reserve-before-loop, 4x use-auto, 2x
    widening cast, 3x unused using-decls removed (grep-verified),
    isolate-declaration, unused fs alias, SaveCompression.hpp NSDMIs.
  - RenderBucket.hpp: 3x return-braced-init; test_restir_di: 2x
    isolate-declaration; test_volumetric_fog_froxel: cert-flp30-c
    int-induction (bit-identical 10 sample points verified);
    test_lighting_clusters_dispatch: use-auto;
    hello_d3d12_triangle: ::atoi -> ::strtol (cert-err34-c).

.clang-tidy: readability-math-missing-parentheses DISABLED with
rationale — `a + b * c` is the standard idiom in a graphics engine;
the rule fired on ~20 correct sites in one sweep and catches no
real defect class for us.

KNOWN ISSUE (separate phase): cd_test_job_graph deadlocked 107 min
under ctest -j8 (2nd hang tonight; passes standalone in 0.15 s).
Troubleshooter RCA in progress; ctest lacks per-test TIMEOUT
properties — hardening queued.

Build: ninja-debug clean.
Tests: 261/261 PASS (full suite, clean re-run).

+++ engine/asset/tests/image/test_bc7.cpp
@@ -62,7 +62,7 @@ TEST(Bc7, NonMultipleOf4DimensionsRoundUp)
 TEST(Bc7, FastVsHighQualityProducesDifferentOutput)
 {
     // Use a gradient so the encoder has interesting work to do.
-    std::vector<std::uint8_t> rgba(16U * 16U * 4U);
+    std::vector<std::uint8_t> rgba(static_cast<std::size_t>(16U) * 16U * 4U);
     for (std::uint32_t y = 0; y < 16; ++y)
     {
         for (std::uint32_t x = 0; x < 16; ++x)
